Fixed some call back parameter bugs
Callback parameters  were incorrect for two reasons.
The first is the shift of the first half of the 64-bit parameter was
only 16 bits in a number of areas.
The second is that a C# casting problem was ANDing the two 32-bits into
one 32-bit instead of creating a 64-bit parameter.

+++ MicroFrameworkPK_v4_3/Samraksh/Interops/Samraksh_SPOT_Hardware/Samraksh_SPOT_Hardware/AnalogInput.cs
@@ -193,7 +193,7 @@ namespace Samraksh.SPOT.Hardware.EmoteDotNow
         /// <param name="time"></param>
         static public void InternalCallback(uint data1, uint data2, DateTime time)
         {
-            MyCallback((data1 << 32) | data2);
+            MyCallback((long)(((long)data1 << 32) | (long)data2));
         }
     }
